+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 The sections should follow the order `Packaging`, `Added`, `Changed`, `Fixed` and `Removed`.
 
 ## [Unreleased]
+## [0.2.4] - 2020-09-02
 ### Added
 - Ajour checks itself for updates (by [tarkah](https://github.com/tarkah))
 - Tukui now handle classic flavor